+++ src/execution.ts
@@ -3,7 +3,7 @@ import * as fs from "fs";
 import * as os from "os";
 import * as path from "path";
 
-import { parseLanguage, Language } from "./language";
+import { parseLanguage, Language, printExpression } from "./language";
 
 function splitAndRemoveEmptyLines(text: string): Array<string> {
   return text
@@ -35,15 +35,17 @@ export async function prepareCode(
       lastLine = lastLine.substring(0, commentIdx).trimEnd();
     }
 
-    const wrappedLastLine = language.logCommand(lastLine);
-    const wrappedCode = [...lines, wrappedLastLine].join("\n");
+    const wrappedLastLine = printExpression(language.output, lastLine);
+    const exit = language.exit ?? '';
+    const wrappedCode = [...lines, wrappedLastLine, exit].join("\n").trim();
     resolve([language, wrappedCode]);
   });
 }
 
 async function doExecuteCode(
   code: string,
-  language: Language
+  language: Language,
+  resultAsComment: boolean
 ): Promise<string> {
   return new Promise(async (resolve, reject) => {
     const tempDir = fs.mkdtempSync(path.join(os.tmpdir(), "vscode-ext-"));
@@ -51,20 +53,21 @@ async function doExecuteCode(
 
     fs.writeFileSync(tempFile, code);
 
-    const commandLine = language.command + " " + tempFile;
+    const commandLine = language.executable + " " + tempFile;
     cp.exec(commandLine, (error, stdout, stderr) => {
       let result: string;
       if (error) {
         result = error.message;
       } else {
         result = stdout || stderr;
-        fs.rmSync(tempFile, { force: true });
-        fs.rmdirSync(tempDir, { recursive: true });
+        fs.rmSync(tempDir, { recursive: true }) 
       }
-      const lines = splitAndRemoveEmptyLines(result);
-      result =
-        (lines.length > 1 ? "\n" : " ") +
-        lines.map((l) => language.comment + " " + l).join("\n");
+      if (resultAsComment) {
+        const lines = splitAndRemoveEmptyLines(result);
+        result =
+          (lines.length > 1 ? "\n" : " ") +
+          lines.map((l) => language.comment + " " + l).join("\n");
+      }
       resolve(result);
     });
   });
@@ -72,8 +75,9 @@ async function doExecuteCode(
 
 export async function executeCode(
   code: string,
-  languageId: string
+  languageId: string,
+  resultAsComment: boolean
 ): Promise<string> {
   const [language, preppedCode] = await prepareCode(code, languageId);
-  return await doExecuteCode(preppedCode, language);
+  return await doExecuteCode(preppedCode, language, resultAsComment);
 }