+++ src/extension.ts
@@ -1,35 +1,47 @@
 import * as vscode from 'vscode';
 import { executeCode } from './execution';
+import { loadLanguages } from './language';
+
+function getCode(editor: vscode.TextEditor) : string {
+  const document = editor.document;
+  const selection = editor.selection;
+  if (selection.isEmpty) {
+    const startPos = new vscode.Position(0, 0);
+    const endOfLine = document.lineAt(selection.active.line).range.end;
+    return document.getText(new vscode.Range(startPos, endOfLine));
+  } else {
+    return document.getText(selection);
+  }
+
+}
+
 export function activate(context: vscode.ExtensionContext) {
   let disposable = vscode.commands.registerCommand('venkat.run', async () => {
     const editor = vscode.window.activeTextEditor;
     if (!editor) {
       return;
-    }
+    }    
+    const code = getCode(editor);
 
     const document = editor.document;
-    const selection = editor.selection;
-    const startPos = new vscode.Position(0, 0);
-    const position = editor.selection.active;
-    const endOfLine = document.lineAt(position.line).range.end;
-
-    const codeRange = new vscode.Range(startPos, endOfLine);
-    const code = document.getText(codeRange);
-
-
     const languageId = document.languageId;
     try {
-      const result = await executeCode(code, languageId);
+      const config = vscode.workspace.getConfiguration('venkat');
+      const resultAsComment = config.get<boolean>('resultAsComment') ?? true;
+      loadLanguages(config.get<{[key:string]:{[key:string]:string}}>('languages'));
+
+      const result = await executeCode(code, languageId, resultAsComment);
 
       if (result) {
-        const position = document.lineAt(editor.selection.active.line).range.end;
+        if (resultAsComment) {          
+          const position = document.lineAt(editor.selection.end).range.end;
+          editor.edit((editBuilder) => editBuilder.insert(position, result));
 
-        editor.edit((editBuilder) => {
-          editBuilder.insert(position, result);
-        });
-
-        const newSelection = new vscode.Selection(position, position.translate(0, (result).length));
-        editor.selection = newSelection;
+          const newSelection = new vscode.Selection(position, position.translate(0, (result).length));
+          editor.selection = newSelection;
+        } else {
+          vscode.window.showInformationMessage(result);
+        }
       }
     } catch (err) {
         const message : string = (err instanceof Error) ? err.message : new String(err).toString();